# api/swift/Sources/MetagenAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metagen Backend API",
    "description": "HTTP API for Metagen superintelligent personal agent",
    "version": "2025.08.19.152833"
  },
  "paths": {
    "/api/chat/approval-response": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Handle Approval Response",
        "description": "Handle tool approval response.",
        "operationId": "handle_approval_response_api_chat_approval_response_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalResponseMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Stream",
        "description": "Stream chat responses as they are generated.",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApprovalDecision": {
        "type": "string",
        "enum": [
          "approved",
          "rejected"
        ],
        "title": "ApprovalDecision",
        "description": "Tool approval decisions."
      },
      "ApprovalResponse": {
        "properties": {
          "tool_id": {
            "type": "string",
            "title": "Tool Id",
            "description": "ID of the tool that was approved/rejected"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision",
            "description": "The approval decision that was processed"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Optional status message",
            "default": "Approval processed successfully"
          }
        },
        "type": "object",
        "required": [
          "tool_id",
          "decision"
        ],
        "title": "ApprovalResponse",
        "description": "Response from the approval endpoint."
      },
      "ApprovalResponseMessage": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MessageType",
            "default": "approval_response"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "default": "METAGEN"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "tool_id": {
            "type": "string",
            "title": "Tool Id"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "tool_id",
          "decision"
        ],
        "title": "ApprovalResponseMessage",
        "description": "User responds to approval request."
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/UserMessage"
              },
              {
                "$ref": "#/components/schemas/ApprovalResponseMessage"
              }
            ],
            "title": "Message",
            "description": "Message to send to the agent - can be a string or a Message object"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session identifier for request routing"
          }
        },
        "type": "object",
        "required": [
          "message",
          "session_id"
        ],
        "title": "ChatRequest",
        "description": "Request to send a message to the agent."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageType": {
        "type": "string",
        "enum": [
          "user",
          "agent",
          "system",
          "thinking",
          "tool_call",
          "approval_request",
          "approval_response",
          "tool_started",
          "tool_result",
          "tool_error",
          "usage",
          "error"
        ],
        "title": "MessageType",
        "description": "Types of messages in the system."
      },
...
      "UserMessage": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MessageType",
            "default": "user"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "default": "METAGEN"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "content"
        ],
        "title": "UserMessage",
        "description": "User chat message."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}