# api/ts/generated/models/ApprovalDecision.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Tool approval decisions.
 */
export enum ApprovalDecision {
    APPROVED = 'approved',
    REJECTED = 'rejected',
}