# api/ts/generated/models/HTTPValidationError.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ValidationError } from './ValidationError.js';
export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};